// BACKENDBOTPROJECT/src/transactions/enums/transaction-detail.enum.ts
export enum TransactionDetail {
  // Memberships
  MEMBERSHIP_WEEKLY = 'MEMBERSHIP_WEEKLY',
  MEMBERSHIP_BASIC = 'MEMBERSHIP_BASIC',
  MEMBERSHIP_AMATEUR = 'MEMBERSHIP_AMATEUR',
  MEMBERSHIP_PRO = 'MEMBERSHIP_PRO',
  MEMBERSHIP_EXPERT = 'MEMBERSHIP_EXPERT',
  MEMBERSHIP_ELITE = 'MEMBERSHIP_ELITE',

  // Operations
  WITHDRAW = 'WITHDRAW',
  DEPOSIT = 'DEPOSIT',
  NETWORK_COMMISSION = 'NETWORK_COMMISSION',
  NETWORK_COMMISSION_RENEWAL = 'NETWORK_COMMISSION_RENEWAL',
  USER_TRANSFER = 'USER_TRANSFER',
  PLAN_PURCHASE = 'PLAN_PURCHASE',
}